nexus: add protective MBR when writing out GPT labels
When testing against 3 nvmf targets I noticed that unlike on my
test machine, the partitions where not scanned at all, if there
was no (protective) MBR.

This commit adds a protective MBR and testing shows that the
partitions automatically show up on that machine as well
